Rename foodNameMiniSearch for clear intent
